Don't crash on anonymous warning
If we see an io.EOF error when unmarshalling the OServiceEvilNotification SNAC, it's likely due to it being
an OServiceEvilNotificationAnon which does not contain any TLVs such as ScreenName (only NewEvil.)

We can safely catch this error and move on. There will be never be a chat context for a blank ScreenName.

Closes #1